Return type should be int
... since that what we store and return.

(generates warnings sometimes if we claim to return a float).

Diffs=
fa8962f39 Return type shold be int

+++ include/rive/animation/linear_animation_instance.hpp
@@ -35,7 +35,7 @@ namespace rive {
         float time() const { return m_Time; }
 
         // Returns the direction that we are currently playing in
-        float direction() const { return m_Direction; }
+        int direction() const { return m_Direction; }
 
         // Update the direction of the animation instance, positive value for
         // forwards Negative for backwards
